fix(client): 荣誉 typeName 反查改 int64 比较消除位宽截断
honor.go 反查曾用 opt.Value == int(typeID)，typeId 超 2^31 时 32 位编译目标静默截断为低位值导致反查必不命中、退化为空 typeName 放行。统一提升为 int64 比较（typeId 是 JSON number 域，int64 才是其真实宽度）。新增 TestAddHonor_TypeNameLookup_LargeTypeID_NoTruncation 以 9999999999 大 id 锁定命中补全行为。

+++ pkg/client/honor.go
@@ -160,7 +160,10 @@ func (c *Client) ensureHonorTypeName(ctx context.Context, token string, typeID i
 		return "", fmt.Errorf("自动反查 typeName 失败: %w", err)
 	}
 	for _, opt := range opts {
-		if opt.Value == int(typeID) {
+		// int64 比较：typeId 是平台 id 域（JSON number），int 转换在 32 位编译目标
+		// 会静默截断高位导致反查必不命中；HonorSelectOption.Value 同为 int 建模，
+		// 比较时统一提升到 int64。
+		if int64(opt.Value) == typeID {
 			c.logDebug("ensureHonorTypeName 自动补全 typeName: typeId=%d → %q", typeID, opt.Label)
 			return opt.Label, nil
 		}
